requirements.txt
streamlit==1.37.1
google-cloud-aiplatform==1.66.0
python-dotenv==1.0.1
google-cloud-firestore==2.16.0
